kafka: Allow multiple in-flight messages
This change allows kafka_connection to be used in non-sequential way,
allowing to "queue" multiple messages without waiting for the previous
to have ended.

Semaphores with count=1 are used to implement queuing of sends and
receives, taking advantage of semaphore FIFO functionality. It is
crucial to preserve the ordering, so that receive reads the correct
response.